requirements.txt
appdirs==1.4.4
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
decorator==4.4.2
emoji==1.7.0
ftfy==6.2.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.5.1
h11==0.14.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
iniconfig==2.0.0
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
playwright==1.42.0
pluggy==1.4.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyee==11.0.1
PySocks==1.7.1
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-slugify==8.0.4
pytube==15.0.0
requests==2.31.0
rsa==4.9
sniffio==1.3.1
sortedcontainers==2.4.0
text-unidecode==1.3
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
Unidecode==1.3.8
update-checker==0.18.0
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==1.7.0
websockets==10.4
wsproto==1.2.0
zipp==3.18.1
